backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
PyJWT==2.8.0
aiohttp==3.8.4